Add components folder
Fix of the violation of code quality guidelines.
There was no components folder in the first
place, so all components have been generated
in the src folder, which is very unclean. It is now
on @sweggerat to fix the paths.
