+++ src/components/molecules/Calendar2D/calendar.scss
@@ -5,7 +5,7 @@
   }
 }
 
-/**
+/** 
  * SPDX-License-Identifier: (EUPL-1.2)
  * Copyright © 2019 Werbeagentur Christian Aichner
  */
