+++ src/components/molecules/Socialdata/socialdata.scss
@@ -16,7 +16,7 @@
   }
 }
 
-/**
+/** 
  * SPDX-License-Identifier: (EUPL-1.2)
  * Copyright © 2019 Werbeagentur Christian Aichner
  */
