+++ src/components/molecules/index.js
@@ -6,8 +6,3 @@ import Socialdata from "./Socialdata";
 import Calendar2D from "./Calendar2D";
 
 export { TabContainer, Avatar, Socialdata, Calendar2D };
-
-/**
- * SPDX-License-Identifier: (EUPL-1.2)
- * Copyright © 2019 Werbeagentur Christian Aichner
- */
