--- src/organisms/tabs/index.js
@@ -1,18 +1,0 @@
-//> Components
-// Import all components to export them for easy access from parent components
-import ResumeTab from './Resume';
-import ProjectsTab from './Projects';
-import OverviewTab from './Overview';
-import EducationTab from './Education';
-
-export {
-    ResumeTab,
-    ProjectsTab,
-    OverviewTab,
-    EducationTab,
-};
-
-/** 
- * SPDX-License-Identifier: (EUPL-1.2)
- * Copyright © 2019 Werbeagentur Christian Aichner
- */
